mlModel/requirements.txt
Flask
Flask-Cors
scipy
tensorflow
keras-facenet
numpy
Pillow
gunicorn
